KAFKA-6388; Recover from rolling an empty segment that already exists (#5986)
There were several reported incidents where the log is rolled to a new segment with the same base offset as an active segment, causing KafkaException: Trying to roll a new log segment for topic partition X-N with start offset M while it already exists. In the cases we have seen, this happens to an empty log segment where there is long idle time before the next append and somehow we get to a state where offsetIndex.isFull() returns true due to _maxEntries == 0. This PR recovers from this state by deleting and recreating the segment and all of its associated index files.

Change-Id: I5682245cef12ca5aa994e37eecc256587cd60ba2
Reviewers: Jason Gustafson <jason@confluent.io>